Update twitch-nonogram-canvas.user.js
Changed FPS of the canvas. There is no reason to have it at 30/60 FPS, since it is mostly static minigame. 
I set it at 1 fps, which might delay slighly the update of  nonogram when playing with more users... This will also let you activate the Sharpness function with lower impact on your CPU :)