Update twitch-nonogram-canvas.user.js
Added visual queues to activity coupon button.
Expiration in less than  30 minutes -> red
Expiration in less than 15 minutes -> deep red. 

In case you try to send a command without a valid coupon (redeemed through the userscript), it will first redeem a coupon. 
I plan to change this to be a toggleable option in the next update.. so you do not think I am enforcing a coupon redemption in case you did it manually.