Update twitch-nonogram-canvas.user.js
Toggleable option for coupo ncheck (guard exports)
If active, it will check if you redeemed a coupon in the last 1 hour. If you did not, it will redeem one and wait around 10 seconds for the bot to process it before sending the command. 
If you have the coupon previously redeemed through the userscript, it will work as before.